PyPI!
Add collections, base file with nothing really in it, need to figure out how
  I want the interface to look.
Move basic setup functions for the tests into baseSetup.py to make collection
  testing easier when I get that far
